+++ apps/common-app/src/examples/RuntimeTests/ReJest/RuntimeTestsApi.ts
@@ -4,13 +4,11 @@ import type { TestComponent } from './TestComponent';
 import type { SharedValue } from 'react-native-reanimated';
 import type { TestConfiguration, TestValue, NullableTestValue, BuildFunction } from './types';
 import { DescribeDecorator, TestDecorator } from './types';
-import {
-  mockWindowDimensions as _mockWindowDimensions,
-  unmockWindowDimensions as _unmockWindowDimensions,
-} from './TestRunner/mockWindowDimensions';
+
 export { Presets } from './Presets';
 
 const testRunner = new TestRunner();
+const windowDimensionsMocker = testRunner.getWindowDimensionsMocker();
 
 type DescribeFunction = (name: string, buildSuite: BuildFunction) => void;
 type TestFunction = (name: string, buildTest: BuildFunction) => void;
@@ -188,11 +186,11 @@ export async function unmockAnimationTimer() {
 }
 
 export async function mockWindowDimensions() {
-  await _mockWindowDimensions();
+  await windowDimensionsMocker.mockWindowDimensions();
 }
 
 export async function unmockWindowDimensions() {
-  await _unmockWindowDimensions();
+  await windowDimensionsMocker.unmockWindowDimensions();
 }
 
 export async function setAnimationTimestamp(timestamp: number) {
